Test - splitting competence
